Add all deprecated method in the package with new name 'org.opensearch.action.support.clustermanager'
