+++ server/src/main/java/org/opensearch/action/support/clustermanager/ClusterManagerNodeOperationRequestBuilder.java
@@ -65,6 +65,17 @@ public abstract class ClusterManagerNodeOperationRequestBuilder<
 
     /**
      * Sets the cluster-manager node timeout in case the cluster-manager has not yet been discovered.
+     *
+     * @deprecated As of 2.1, because supporting inclusive language, replaced by {@link #setClusterManagerNodeTimeout(TimeValue)}
+     */
+    @SuppressWarnings("unchecked")
+    @Deprecated
+    public final RequestBuilder setMasterNodeTimeout(TimeValue timeout) {
+        return setClusterManagerNodeTimeout(timeout);
+    }
+
+    /**
+     * Sets the cluster-manager node timeout in case the cluster-manager has not yet been discovered.
      */
     @SuppressWarnings("unchecked")
     public final RequestBuilder setClusterManagerNodeTimeout(String timeout) {
@@ -72,4 +83,14 @@ public abstract class ClusterManagerNodeOperationRequestBuilder<
         return (RequestBuilder) this;
     }
 
+    /**
+     * Sets the cluster-manager node timeout in case the cluster-manager has not yet been discovered.
+     *
+     * @deprecated As of 2.1, because supporting inclusive language, replaced by {@link #setClusterManagerNodeTimeout(String)}
+     */
+    @SuppressWarnings("unchecked")
+    @Deprecated
+    public final RequestBuilder setMasterNodeTimeout(String timeout) {
+        return setClusterManagerNodeTimeout(timeout);
+    }
 }
