+++ server/src/main/java/org/opensearch/action/support/clustermanager/ClusterManagerNodeRequest.java
@@ -48,7 +48,15 @@ public abstract class ClusterManagerNodeRequest<Request extends ClusterManagerNo
 
     public static final TimeValue DEFAULT_CLUSTER_MANAGER_NODE_TIMEOUT = TimeValue.timeValueSeconds(30);
 
+    /** @deprecated As of 2.1, because supporting inclusive language, replaced by {@link #DEFAULT_CLUSTER_MANAGER_NODE_TIMEOUT} */
+    @Deprecated
+    public static final TimeValue DEFAULT_MASTER_NODE_TIMEOUT = DEFAULT_CLUSTER_MANAGER_NODE_TIMEOUT;
+
     protected TimeValue clusterManagerNodeTimeout = DEFAULT_CLUSTER_MANAGER_NODE_TIMEOUT;
+
+    /** @deprecated As of 2.1, because supporting inclusive language, replaced by {@link #clusterManagerNodeTimeout} */
+    @Deprecated
+    protected TimeValue masterNodeTimeout = clusterManagerNodeTimeout;
 
     protected ClusterManagerNodeRequest() {}
 
@@ -74,6 +82,17 @@ public abstract class ClusterManagerNodeRequest<Request extends ClusterManagerNo
 
     /**
      * A timeout value in case the cluster-manager has not been discovered yet or disconnected.
+     *
+     * @deprecated As of 2.1, because supporting inclusive language, replaced by {@link #clusterManagerNodeTimeout(TimeValue)}
+     */
+    @SuppressWarnings("unchecked")
+    @Deprecated
+    public final Request masterNodeTimeout(TimeValue timeout) {
+        return clusterManagerNodeTimeout(timeout);
+    }
+
+    /**
+     * A timeout value in case the cluster-manager has not been discovered yet or disconnected.
      */
     public final Request clusterManagerNodeTimeout(String timeout) {
         return clusterManagerNodeTimeout(
@@ -81,11 +100,20 @@ public abstract class ClusterManagerNodeRequest<Request extends ClusterManagerNo
         );
     }
 
+    /**
+     * A timeout value in case the cluster-manager has not been discovered yet or disconnected.
+     *
+     * @deprecated As of 2.1, because supporting inclusive language, replaced by {@link #clusterManagerNodeTimeout(String)}
+     */
+    @Deprecated
+    public final Request masterNodeTimeout(String timeout) {
+        return clusterManagerNodeTimeout(timeout);
+    }
+
     public final TimeValue clusterManagerNodeTimeout() {
         return this.clusterManagerNodeTimeout;
     }
 
-    // Preserve the method so that classes implements AcknowledgedRequest don't need to override it for backwards compatibility.
     /** @deprecated As of 2.1, because supporting inclusive language, replaced by {@link #clusterManagerNodeTimeout()} */
     @Deprecated
     public final TimeValue masterNodeTimeout() {
