+++ server/src/main/java/org/opensearch/action/support/master/MasterNodeOperationRequestBuilder.java
@@ -37,7 +37,6 @@ import org.opensearch.action.ActionResponse;
 import org.opensearch.action.support.clustermanager.ClusterManagerNodeOperationRequestBuilder;
 import org.opensearch.action.support.clustermanager.ClusterManagerNodeRequest;
 import org.opensearch.client.OpenSearchClient;
-import org.opensearch.common.unit.TimeValue;
 
 /**
  * Base request builder for cluster-manager node operations
@@ -54,23 +53,5 @@ public abstract class MasterNodeOperationRequestBuilder<
 
     protected MasterNodeOperationRequestBuilder(OpenSearchClient client, ActionType<Response> action, Request request) {
         super(client, action, request);
-    }
-
-    /**
-     * Sets the cluster-manager node timeout in case the cluster-manager has not yet been discovered.
-     */
-    @SuppressWarnings("unchecked")
-    @Deprecated
-    public final RequestBuilder setMasterNodeTimeout(TimeValue timeout) {
-        return setClusterManagerNodeTimeout(timeout);
-    }
-
-    /**
-     * Sets the cluster-manager node timeout in case the cluster-manager has not yet been discovered.
-     */
-    @SuppressWarnings("unchecked")
-    @Deprecated
-    public final RequestBuilder setMasterNodeTimeout(String timeout) {
-        return setClusterManagerNodeTimeout(timeout);
     }
 }
